refactor(server): move pure naming rules out of the Disk namespace
`workspace/disk/naming.ts` held no filesystem access — only string
rules — while its path asserted a substrate. Every consumer that needed
a safe filename or a comparison key had to import through a module
named `disk` to get logic that has nothing to do with disks.

Move the rules to `utils/naming.ts` and update all 15 call sites. The
module is deleted outright rather than left as a forwarding shim: every
importer is updated in the same commit, so a shim would only preserve
the wrong direction for the next person to follow.

Logic is unchanged. The file is byte-identical to the extraction Phase 5
performs as a side effect of needing the rules in a second backend, so
that branch rebases onto this with the file already in place.

Phase 4.5, step 1 of 6 (proposal §12.5.5).
